Move skilling pet dry streaks into patron drystreak
